requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
anthropic==0.25.0
